# swagger.yaml
openapi: 3.0.3
info:
  title: Reminder API
  description: Medication and reminder organizer API
  version: 1.0.0
...
paths:
  /reminders/:
    post:
      summary: Create a reminder
      tags:
        - Reminders
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  minLength: 1
                description:
                  type: string
                dueDate:
                  type: string
                  format: date-time
              required:
                - title
                - dueDate
              additionalProperties: false
      responses:
        "200":
          description: Default Response
    get:
      summary: List all reminders
      tags:
        - Reminders
      responses:
        "200":
          description: Default Response
  /reminders/{id}:
    put:
      tags:
        - Reminders
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: string
                  minLength: 1
                title:
                  type: string
                  minLength: 1
                description:
                  type: string
                dueDate:
                  type: string
                  format: date-time
                completed:
                  type: boolean
              additionalProperties: false
      parameters:
        - schema:
            type: string
          in: path
          name: id
          required: true
      responses:
        "200":
          description: Default Response
...
servers:
  - url: http://localhost:3333
    description: Development